getTimegrid fixed wrong days (#6)
getTimegrid previously incorrectly mapped the timegrid data onto the weekdays. This commit addresses the issue by switching the indices in the TimeGrid._fromList() constructor. 
As none of the contributors can currently test the behaviour for Saturday and Sunday, it remains to be seen if there will be any problems in the future.